flow: turn flags field into uint64_t
Entire space is used.

(cherry picked from commit a2e9ac21f845e94b7e4b16a0cf821bb90e9094b4)

+++ src/flow.c
@@ -1226,6 +1226,7 @@ uint16_t FlowGetDestinationPort(Flow *flow)
 {
     return flow->dp;
 }
+
 /**
  * \brief Get flow flags.
  *
@@ -1233,7 +1234,7 @@ uint16_t FlowGetDestinationPort(Flow *flow)
  * opaque pointer to the flow structure.
  */
 
-uint32_t FlowGetFlags(Flow *flow)
+uint64_t FlowGetFlags(Flow *flow)
 {
     return flow->flags;
 }
